Automate_email/requirements.txt
pandas==2.1.4
openai==1.12.0
python-dotenv==1.0.0 